Phase-B preflight fixes: soak-gate Windows RSS, regression_e2e vendored-crates path, clippy dead-code/conversion, cargo fmt
1. soak_gate.py: Add Windows GetProcessMemoryInfo fallback for sample_rss_bytes().
   ROOT CAUSE: On Windows, /proc/self/status doesn't exist and the 'resource'
   module is absent (POSIX-only). sample_rss_bytes() returned 0, making the
   gate VACUOUS — it could not detect any leak, so test_leaky_workload_is_caught
   failed (the leaky workload passed). Fix: ctypes + psapi.GetProcessMemoryInfo
   with proper argtypes/restype to read WorkingSetSize (current RSS).

2. regression_e2e.py: Fix vendored-crates path failure for capture suites.
   ROOT CAUSE: The script hardcoded --offline for capture suite tests, but
   the vendor dir (/tmp/vw/vendor) doesn't exist on this box. With --offline
   and no vendor, cargo cannot resolve the dependency tree. Additionally,
   ensure_vendored_config() wrote a .cargo/config.toml pointing at the
   missing vendor dir, compounding the failure. Fix: only use --offline when
   the vendor dir exists; skip vendored config when vendor is absent.
   Also: Windows path-length workaround for cargo fmt --all --check (OS error
   206 / ERROR_FILENAME_EXCED_RANGE) — falls back to per-crate fmt checks.

3. phase_b_preflight.py: Same Windows fmt workaround for row 6.

4. Capture suite Cargo.lock files: ring 0.17.16000 → 0.17.14 (the pinned
   version never existed on crates.io; cargo update resolved it).

5. Clippy fixes: dead_code #[expect] on unused test constants, useless u64
   conversion removal, manual checked_div replacement.

6. cargo fmt: applied to 4 crates with formatting diffs.

NOT a hardware threshold issue. NOT a code defect in the soak-gate logic.
The soak-gate's trend decision (analyze_trend) is a pure function that was
always correct — the failure was in the RSS sampling layer (platform
portability gap), not the threshold logic. The 4-of-12 memory channel
configuration is irrelevant to this fix.

+++ rust/crates/pump-quant-app/tests/one_authority_laws.rs
@@ -180,7 +180,7 @@ fn the_gate_and_the_curve_are_one_impact_model() {
     ] {
         let den = cost_model::impact_den_for(vsol);
         let gate_bps = CLIP / den; // ImpactCurve::linear_test semantics
-        let curve_bps = u64::from(pump_quant_app::curve_fill::own_impact_bps(vsol, CLIP).unwrap());
+        let curve_bps = pump_quant_app::curve_fill::own_impact_bps(vsol, CLIP).unwrap();
         assert!(
             gate_bps.abs_diff(curve_bps) <= 1,
             "the gate ({gate_bps} bps) and the curve ({curve_bps} bps) must price our own \
